+++ exchangelib/configuration.py
@@ -31,8 +31,7 @@ class Configuration(object):
         account = Account(primary_smtp_address='john@example.com', credentials=credentials, autodiscover=True)
 
     """
-    def __init__(self, credentials, server=None, has_ssl=True, service_endpoint=None, auth_type=None,
-                 verify_ssl=True, version=None):
+    def __init__(self, credentials, server=None, has_ssl=True, service_endpoint=None, auth_type=None, version=None):
         if auth_type is not None and auth_type not in AUTH_TYPE_MAP:
             raise ValueError('Unsupported auth type %s' % auth_type)
         if not (server or service_endpoint):
@@ -44,7 +43,6 @@ class Configuration(object):
             service_endpoint=service_endpoint,
             auth_type=auth_type,
             credentials=credentials,
-            verify_ssl=verify_ssl,
             version=version
         )